app/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
python-dotenv==1.0.0
PyPDF2==3.0.1
spacy==3.5.0
requests==2.31.0
python-dateutil==2.8.2